Main Page Design Revision
There is a lot of opportunity to change styles, there are styles that are repeated and methods that are not necessary.

For now, an attempt was made to make the cards look the closest to Figma.

This point is of consideration because being redundant makes us lose efficiency.

So the changes that were made are:
1. Improve the scrollbar
2. Change the design of the Cards
3. Add Bebas Neue

P.S. The images look strange, we will need to put that the main image has to be necessarily 1: 1 ratio